python_ocr/requirements.txt
flask
opencv-python-headless
numpy
pillow
requests
google-cloud-vision
